requirements.txt
django-mptt>=0.6
